testing scipy solvers
